Add defichain data fetching feature
This commit adds feature to fetch historical data from the DefiChain Data API. It includes new dependencies in poetry.lock for packages such as pandas and tqdm required for handling and presenting the fetched data. It also includes implementation of pytest for testing the feature. This feature can fetch history data such as daily, hourly or minutely by providing appropriate parameters. It also provides the flexibility to mention a specific time frame for the data by providing 'from_timestamp' or 'to_timestamp'.